requirements.txt
python-telegram-bot==20.6
flask[async]
gunicorn
gspread
oauth2client
python-dotenv